Extract tests related to metrics consumers to a new class.
